Fixing an issue with Pubmed entries without abstracts
Any Pubmed entries that is lacking an abstract text causes an error
during the parsing of the XML file. As example use this article:

http://www.ncbi.nlm.nih.gov/pubmed/22442467

You can download its XML file with

$ curl "http://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=pubmed&id=$22442467&retmode=xml" > 22442467.xml

The is no "Abstract" and "AbstractText" element in there. The
modification circumvent the error and an empty string is returned by
the modified function.